requirements.txt
pyformlang
antlr4-python3-runtime
graphviz